🐛 fix: add useLogout hook & unknown user is logged out
